PyCode/aeo report/requirements.txt
# AEO Report Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
advertools>=0.14.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0

# Web scraping and analysis
requests>=2.28.0
beautifulsoup4>=4.11.0
playwright>=1.30.0

# Network analysis
networkx>=2.8.0

# Web server
Flask>=2.2.0

# Environment management
python-dotenv>=1.0.0

# Generative AI
openai>=1.3.0

# Database
psycopg2-binary>=2.9.0

# Logging and utilities
# Use Python's built-in 'logging' module; do NOT add PyPI 'logging' package (it is obsolete and breaks installs)

# Optional: For enhanced PDF generation (if needed)
# weasyprint>=57.0
# reportlab>=3.6.0